Move number input to its own folder for work on spinner buttons
